Remove area processing functions
